# docs/openapi.yaml
openapi: 3.0.3
info:
  title: El Inmortal 2 API
  version: 1.0.0
  description: API v1 with API key auth for MCP automation and artist operations.
servers:
  - url: http://localhost:3000/api/v1
security:
  - ApiKeyAuth: []
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
  schemas:
...
    Album:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        artist:
          type: string
        release_date:
          type: string
          format: date
          nullable: true
        status:
          type: string
        description:
          type: string
          nullable: true
        cover_image_path:
          type: string
          nullable: true
    Track:
      type: object
      properties:
        id:
          type: integer
        track_number:
          type: integer
        title:
          type: string
        album_id:
          type: integer
          nullable: true
        producer_id:
          type: integer
          nullable: true
        status:
          type: string
        track_type:
          type: string
        is_single:
          type: integer
        is_primary:
          type: integer
paths:
...
  /albums:
    get:
      summary: List albums
      responses:
        '200':
          description: albums list
    post:
      summary: Create album
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Album'
      responses:
        '201':
          description: created
...
  /tracks:
    get:
      summary: List tracks
      parameters:
        - in: query
          name: filter
          schema:
            type: string
            enum: [singles, primary, album, pending]
      responses:
        '200':
          description: tracks list
    post:
      summary: Create track
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Track'
      responses:
        '201':
          description: created
...